+++ server/src/commands.ts
@@ -1,24 +1,24 @@
 import { Command } from 'vscode-languageserver';
 
 // Create commands
-const COMMAND_LINT = Command.create('Analyze code', 'groovyLint.lint');
-const COMMAND_LINT_FIX = Command.create('Fix all errors', 'groovyLint.lintFix');
-const COMMAND_LINT_QUICKFIX = Command.create('Quick fix', 'groovyLint.quickFix');
-const COMMAND_LINT_QUICKFIX_FILE = Command.create('Quick fix in entire file', 'groovyLint.quickFixFile');
-const COMMAND_SUPPRESS_WARNING_LINE = Command.create('Ignore this error', 'groovyLint.addSuppressWarning');
-const COMMAND_SUPPRESS_WARNING_FILE = Command.create('Ignore this error in entire file', 'groovyLint.addSuppressWarningFile');
-const COMMAND_IGNORE_ERROR_FOR_ALL_FILES = Command.create('Ignore this error in workspace', 'groovyLint.alwaysIgnoreError');
-const COMMAND_SHOW_RULE_DOCUMENTATION = Command.create('Ignore this error in all files', 'groovyLint.showRuleDocumentation');
-const COMMAND_LINT_FOLDER = Command.create('Lint Groovy in folder', 'groovyLint.lintFolder');
+export const COMMAND_LINT = Command.create('Analyze code', 'groovyLint.lint');
+export const COMMAND_LINT_FIX = Command.create('Fix all auto-fixable problems', 'groovyLint.lintFix');
+export const COMMAND_LINT_QUICKFIX = Command.create('Quick fix this line', 'groovyLint.quickFix');
+export const COMMAND_LINT_QUICKFIX_FILE = Command.create('Quick fix rule in this entire file', 'groovyLint.quickFixFile');
+export const COMMAND_DISABLE_ERROR_FOR_LINE = Command.create('Disable rule for this line', 'groovyLint.disableRule');
+export const COMMAND_DISABLE_ERROR_FOR_FILE = Command.create('Disable rule for this entire file', 'groovyLint.disableRuleInFile');
+export const COMMAND_DISABLE_ERROR_FOR_PROJECT = Command.create('Disable rule for this entire project', 'groovyLint.disableRuleInProject');
+export const COMMAND_SHOW_RULE_DOCUMENTATION = Command.create('Show documentation for rule', 'groovyLint.showDocumentationForRule');
+export const COMMAND_LINT_FOLDER = Command.create('Lint Groovy file in this folder', 'groovyLint.lintFolder');
 
 export const commands = [
 	COMMAND_LINT,
 	COMMAND_LINT_FIX,
 	COMMAND_LINT_QUICKFIX,
 	COMMAND_LINT_QUICKFIX_FILE,
-	COMMAND_SUPPRESS_WARNING_LINE,
-	COMMAND_SUPPRESS_WARNING_FILE,
-	COMMAND_IGNORE_ERROR_FOR_ALL_FILES,
+	COMMAND_DISABLE_ERROR_FOR_LINE,
+	COMMAND_DISABLE_ERROR_FOR_FILE,
+	COMMAND_DISABLE_ERROR_FOR_PROJECT,
 	COMMAND_SHOW_RULE_DOCUMENTATION,
 	COMMAND_LINT_FOLDER
 ];